+++ package.json
@@ -17,13 +17,15 @@
 		"@graphql-tools/schema": "^8.3.2",
 		"@kitql/vite-plugin-watch-and-run": "^0.4.2",
 		"@sveltejs/adapter-auto": "^1.0.0",
+		"@sveltejs/adapter-cloudflare": "^1.0.0",
+		"@sveltejs/adapter-cloudflare-workers": "^1.0.6",
 		"@sveltejs/kit": "^1.3.0",
 		"apollo-server": "^3.6.3",
 		"apollo-server-core": "^3.6.3",
 		"concurrently": "^7.0.0",
 		"express": "^4.17.3",
 		"feather-icons": "^4.28.0",
-		"graphql": "15.7.2",
+		"graphql": "^15.8.0",
 		"graphql-subscriptions": "^2.0.0",
 		"graphql-tag": "^2.12.6",
 		"graphql-yoga": "^3.5.1",
